requirements.txt
django==2.2.6
graphene-django==2.6.0
django-cors-headers==3.1.1
requests==2.22.0
django-environ==0.4.5
psycopg2==2.8.4
celery==4.3.0
redis==3.3.11
channels==2.3.1
channels-redis==2.4.1
django-graphql-jwt==0.2.2
django-filter==2.2.0